.history/CellSegmentation/requirements_20240313030315.txt
absl-py==1.4.0

anndata==0.7.5

astunparse==1.6.3

cachetools==5.3.0

charset-normalizer==3.1.0

cycler==0.11.0

flatbuffers==23.3.3

fonttools==4.39.3

gast==0.5.3

google-auth==2.17.1

google-auth-oauthlib==0.4.6

google-pasta==0.2.0

grpcio==1.53.0

h5py==3.8.0

idna==3.4

imagecodecs==2023.3.16

importlib-metadata==6.1.0

joblib==1.2.0

keras==2.8.0

keras-preprocessing==1.1.2

kiwisolver==1.4.4

libclang==16.0.0

llvmlite==0.39.1

markdown==3.4.3

markupsafe==2.1.2

matplotlib==3.5.0

natsort==8.3.1

networkx==3.0

numba==0.56.4

numpy==1.22.4

oauthlib==3.2.2

opt-einsum==3.3.0

packaging==23.0

pandas==1.3.4

patsy==0.5.3

pillow==9.4.0

protobuf==3.19.6

pyasn1==0.4.8

pyasn1-modules==0.2.8

pynndescent==0.5.8

pyparsing==3.0.9

python-dateutil==2.8.2

pytz==2023.3

requests==2.28.2

requests-oauthlib==1.3.1

rsa==4.9

scanpy==1.9.3

scikit-learn==1.2.2

scipy==1.10.1

seaborn==0.12.2

session-info==1.0.0

setuptools-scm==7.1.0

six==1.16.0

statsmodels==0.13.5

stdlib-list==0.8.0

tensorboard==2.8.0

tensorboard-data-server==0.6.1

tensorboard-plugin-wit==1.8.1

tensorflow==2.8.2

tensorflow-addons==0.16.1

tensorflow-estimator==2.8.0

tensorflow-io-gcs-filesystem==0.32.0

termcolor==2.2.0

threadpoolctl==3.1.0

tomli==2.0.1

tqdm==4.65.0

typeguard==3.0.2

typing-extensions==4.5.0

umap-learn==0.5.3

urllib3==1.26.15

werkzeug==2.2.3

wrapt==1.15.0

zipp==3.15.0

spateo-release
